rename into ebrains_seafile
