+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = -s -v --doctest-modules --ignore=build --ignore=dist --ignore=hbp_seafile.egg-info --ignore=setup.py
+addopts = -s -v --doctest-modules --ignore=build --ignore=dist --ignore=ebrains_drive.egg-info --ignore=setup.py
